"Initialize the project structure with Nest.js"
In this commit, the project structure was initialized utilizing the Nest.js framework. Multiple files were created to form a base structure, such as controller, service, and module files for 'store', as well as 'app' related files. Configuration and setup files for TypeScript, Eslint, and Prettier were also included along with logging and dependency management settings in .gitignore and package-lock.json respectively.